initial struct defs and update to ls to display them.
This commit adds the ownerid, groupid and mode attributes into both the
on-disk inode structure and the in-memory inode structure. It also
loads the values into a stat structure that is consumed by the ls
application which displays the new values.

mkfs.c has been updated to add mock values for the attributes for the
time being. Later commits will extend on this base.

Signed-off-by: Jim Huang <jserv.tw@gmail.com>